refactor(snapshot): move text-surface into snapshot presentation
