+++ src/daemon/handlers/interaction-read.ts
@@ -8,7 +8,10 @@ import { runtimeExecutionFromContext } from '../snapshot-runtime-capture-input.t
 import { emitDiagnostic } from '@agent-device/host-kit/diagnostics';
 import type { SessionState } from '../types.ts';
 import type { SnapshotNode } from '@agent-device/kernel/snapshot';
-import { extractReadableText, prefersValueForReadableText } from '../../utils/text-surface.ts';
+import {
+  extractReadableText,
+  prefersValueForReadableText,
+} from '../../snapshot/snapshot-presentation/text-surface.ts';
 import type { ContextFromFlags } from './interaction-common.ts';
 import { resolveRectCenter } from './interaction-targeting.ts';
