+++ src/snapshot/snapshot-presentation/text-surface.ts
